+++ src/main/java/io/github/ianoliveira/ac/firstproj/repositories/VideoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface VideoRepository extends JpaRepository<Video, Long> {
+public interface VideoRepository extends JpaRepository<Video, Integer> {
 
     Optional<Video> findByUrl(String url);
 }
